Isolate database expensive reads in its own module and isolate its `use` lines to avoid unused warnings
